refactor(extractor): moving Perplexity sources extracting logic
