+++ src/scripts/content/extractor/sources/ExtractorSources.js
@@ -1,0 +1,7 @@
+import {Extractor} from "../types/Extractor";
+
+export default class ExtractorSources extends Extractor {
+  async extractSources(content, format) {
+    throw new Error("Not implemented");
+  }
+}
